usual commit

+++ E-Olymp/matchsModel_3.kt
@@ -2,7 +2,6 @@ import java.util.*
 
 fun main(args: Array<String>) {
         val sc = Scanner(System.`in`)
-        var n = sc.nextInt()
-
+        var n = sc.nextLong()
         println("${4 * (2 * n + 1)}")
 }
